fazendo login

+++ AutoSavePI4/Entidades/CAMINHAO.cs
@@ -5,7 +5,7 @@ namespace AutoSavePI4.Entidades
     public class Caminhao
     {
         [Key]
-        public long Codigo { get; set; }
+        public int Codigo { get; set; }
         public string? Placa { get; set; }
         public bool Bitrem { get; set; }
         public string? Carroceria { get; set; }
